requirements.txt
tensorflow>=2.20.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.22.0
fastapi>=0.108.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.29.0
matplotlib>=3.8.0
seaborn>=0.13.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
